+++ packages/kotlin-markdown-core/src/commonMain/kotlin/com/nouprax/markdown/core/visitor/Visitor.kt
@@ -43,6 +43,8 @@ public interface Visitor<Result> {
 
     public fun visitComment(node: Comment): Result
 
+    public fun visitCrossLink(node: CrossLink): Result
+
     public fun visitFormula(node: Formula): Result
 
     public fun visitEmphasis(node: Emphasis): Result